+++ Controls-demo/dateRange/Input/Default/Default.wml
@@ -1,0 +1,11 @@
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__cell">
+        <Controls.dateRange:Input
+                bind:startValue="_startDate"
+                bind:endValue="_endDate"
+                mask="{{_mask}}"
+                startTagStyle="{{'danger'}}"
+                endTagStyle="{{'success'}}"/>
+
+    </div>
+</div>
