+++ Controls/_dateRange/Input.ts
@@ -35,7 +35,7 @@ interface IDateRangeInputOptions extends IDateRangeValidatorsOptions {
  * @mixes Controls/_interface/IDateRangeValidators
  * 
  * @public
- * @demo Controls-demo/Input/Date/RangeInput
+ * @demo Controls-demo/dateRange/Default/Index
  * @author Красильников А.С.
  */
 
@@ -51,7 +51,7 @@ interface IDateRangeInputOptions extends IDateRangeValidatorsOptions {
  *
  * 
  * @public
- * @demo Controls-demo/Input/Date/RangeInput
+ * @demo Controls-demo/dateRange/Default/Index
  * @author Красильников А.С.
  */
 export default class DateRangeInput extends Control<IDateRangeInputOptions> implements
