+++ Controls/_interface/IDateRangeValidators.ts
@@ -32,38 +32,7 @@ export default interface IDateRangeValidators {
  * При вызове валидатора в качестве аргументов он получает введнное в поле значение и заданные из прикладного кода
  * аргументы.
  *
- * @example
- * Поле ввода периода с валидатором, который проверяет, что в поле ввода начала периода введено значение.
- * Так же этот валидатор устанавливается в поле ввода в диалоге выбора периода, который открывается
- * по клику на кнопку рядом с полями ввода.
- *
- * <pre class="brush: html">
- * <Controls.dateRange:Input>
- *    <ws:startValueValidators>
- *       <ws:Array>
- *          <ws:Function>Controls/validate:isRequired</ws:Function>
- *       </ws:Array>
- *    </ws:startValueValidators>
- * </Controls.dateRange:Input>
- * </pre>
- *
- * Поле ввода с прикладным валидатором, который проверяет, что начало периода лежит в интервале между _startValue
- * и _endValue.
- *
- * <pre class="brush: html">
- * <Controls.dateRange:Input>
- *    <ws:startValueValidators>
- *       <ws:Array>
- *          <ws:Object>
- *             <ws:validator>
- *                <ws:Function>Controls-demo/Input/DateBase/Validators/isInRange:default</ws:Function>
- *             </ws:validator>
- *             <ws:arguments startValue="{{_startValue}}" endValue="{{_endValue}}"/>
- *          </ws:Object>
- *       </ws:Array>
- *    </ws:startValueValidators>
- * </Controls.dateRange:Input>
- * </pre>
+ * @demo Controls-demo/dateRange/Validators/Index
  *
  * @see Controls/_interface/IDateRangeValidators#endValueValidators
  * @see Controls/_interface/IDateRangeValidators#validateByFocusOut
@@ -77,38 +46,7 @@ export default interface IDateRangeValidators {
  * При вызове валидатора в качестве аргументов он получает введнное в поле значение и заданные из прикладного кода
  * аргументы.
  *
- * @example
- * Поле ввода периода с валидатором, который проверяет, что в поле ввода коца периода введено значение.
- * Так же этот валидатор устанавливается в поле ввода в диалоге выбора периода, который открывается
- * по клику на кнопку рядом с полями ввода.
- *
- * <pre class="brush: html">
- * <Controls.dateRange:Input>
- *     <ws:endValueValidators>
- *         <ws:Array>
- *             <ws:Function>Controls/validate:isRequired</ws:Function>
- *         </ws:Array>
- *     </ws:endValueValidators>
- * </Controls.dateRange:Input>
- * </pre>
- *
- * Поле ввода с прикладным валидатором, который проверяет, что конец периода лежит в интервале между _startValue
- * и _endValue.
- *
- * <pre class="brush: html">
- * <Controls.dateRange:Input>
- *     <ws:endValueValidators>
- *         <ws:Array>
- *             <ws:Object>
- *                 <ws:validator>
- *                     <ws:Function>Controls-demo/Input/DateBase/Validators/isInRange:default</ws:Function>
- *                 </ws:validator>
- *                 <ws:arguments startValue="{{_startValue}}" endValue="{{_endValue}}"/>
- *             </ws:Object>
- *         </ws:Array>
- *     </ws:endValueValidators>
- * </Controls.dateRange:Input>
- * </pre>
+ * @demo Controls-demo/dateRange/Validators/Index
  *
  * @see Controls/_interface/IDateRangeValidators#startValueValidators
  * @see Controls/_interface/IDateRangeValidators#validateByFocusOut
@@ -120,12 +58,7 @@ export default interface IDateRangeValidators {
  * будет срабатывать только по внешнему триггеру, например при валидации формы.
  * @default true
  *
- * @example
- * Включаем срабатывание валидации только по внешнему тригеру. Например при валидации формы.
- *
- * <pre class="brush: html">
- * <Controls.input:Date validateByFocusOut="{{false}}"/>
- * </pre>
+ * @demo Controls-demo/dateRange/Validators/Index
  *
  * @see Controls/_interface/IDateRangeValidators#startValueValidators
  * @see Controls/_interface/IDateRangeValidators#endValueValidators
